+++ common/src/main/java/com/coolerpromc/uncrafteverything/UncraftEverything.java
@@ -4,10 +4,13 @@ import com.coolerpromc.uncrafteverything.block.UEBlocks;
 import com.coolerpromc.uncrafteverything.blockentity.UEBlockEntities;
 import com.coolerpromc.uncrafteverything.config.FTBQuestProgressionConfig;
 import com.coolerpromc.uncrafteverything.config.PerItemExpCostConfig;
+import com.coolerpromc.uncrafteverything.config.UncraftEverythingConfig;
 import com.coolerpromc.uncrafteverything.item.UECreativeTab;
 import com.coolerpromc.uncrafteverything.screen.UEMenuTypes;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Player;
 
-public class CommonClass {
+public class UncraftEverything {
     public static boolean AUTO_MOVE = true;
 
     public static void init() {
@@ -20,5 +23,11 @@ public class CommonClass {
         FTBQuestProgressionConfig.load();
         PerItemExpCostConfig.startWatcher();
         FTBQuestProgressionConfig.startWatcher();
+    }
+
+    public static void onPlayerLogin(Player player){
+        if (UncraftEverythingConfig.restrictAmbiguouslyCraftedItems()){
+            player.sendSystemMessage(Component.literal("[Uncraft Everything] Restrict Ambiguously Crafted Items config is enabled, recipe that use ItemTags ingredient will not be able uncrafted!"));
+        }
     }
 }