+++ common/src/main/java/com/coolerpromc/uncrafteverything/UncraftEverythingClient.java
@@ -6,7 +6,7 @@ import net.minecraft.world.item.crafting.RecipeHolder;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CommonClientClass {
+public class UncraftEverythingClient {
     public static ClientBoundResponseConfigPayload payloadFromServer;
     public static List<RecipeHolder<?>> recipesFromServer = new ArrayList<>();
 }