+++ neoforge/src/main/java/com/coolerpromc/uncrafteverything/config/NeoForgeUncraftEverythingConfig.java
@@ -26,6 +26,7 @@ public class NeoForgeUncraftEverythingConfig {
     public final ModConfigSpec.BooleanValue onlyAllowDefinedProgression;
     public final ModConfigSpec.BooleanValue outputEnchantedBook;
     public final ModConfigSpec.BooleanValue prioritizeVanillaIngredientRecipe;
+    public final ModConfigSpec.BooleanValue restrictAmbiguouslyCraftedItems;
 
     static {
         Pair<NeoForgeUncraftEverythingConfig, ModConfigSpec> pair = new ModConfigSpec.Builder().configure(NeoForgeUncraftEverythingConfig::new);
@@ -43,7 +44,8 @@ public class NeoForgeUncraftEverythingConfig {
         builder.push("Restrictions");
         restrictionType = builder.comment("The type of restriction to be used.").defineEnum("restrictionType", UncraftEverythingConfig.RestrictionType.BLACKLIST, UncraftEverythingConfig.RestrictionType.values());
         restrictions = builder.comment("A list of items that can/cannot be uncrafted depending on type of restriction.", "Invalid input will cause config reset at runtime.", "Format: modid:item_name / modid:* / modid:*_glass / modid:black_* / modid:red_*_glass / modid:red_*_glass* / #modid:item_tag_name", "Press F3 + h in game and hover item to check their modid:name")
-                .defineList("restrictions", List.of("uncrafteverything:uncrafting_table", "minecraft:crafting_table"), () -> "", o -> o instanceof String && Identifier.tryParse((String) o) != null || o.toString().contains("*") || tryParseTagKey(o.toString().substring(1)).isPresent());
+                .defineList("restrictions", List.of(), () -> "", o -> o instanceof String && Identifier.tryParse((String) o) != null || o.toString().contains("*") || tryParseTagKey(o.toString().substring(1)).isPresent());
+        restrictAmbiguouslyCraftedItems = builder.comment("Restrict recipe that use ItemTags ingredient from uncrafting.").define("restrictAmbiguouslyCraftedItems", false);
         builder.pop();
 
         builder.push("AllowEnchantedItems");
@@ -105,7 +107,8 @@ public class NeoForgeUncraftEverythingConfig {
                 CONFIG.prioritizeVanillaIngredientRecipe.getAsBoolean(),
                 CONFIG.enableProgression.getAsBoolean(),
                 CONFIG.onlyAllowDefinedProgression.getAsBoolean(),
-                CONFIG.outputEnchantedBook.getAsBoolean()
+                CONFIG.outputEnchantedBook.getAsBoolean(),
+                CONFIG.restrictAmbiguouslyCraftedItems.getAsBoolean()
         );
     }
 
@@ -123,6 +126,7 @@ public class NeoForgeUncraftEverythingConfig {
         CONFIG.enableProgression.set(UncraftEverythingConfig.enableProgression);
         CONFIG.onlyAllowDefinedProgression.set(UncraftEverythingConfig.onlyAllowDefinedProgression);
         CONFIG.outputEnchantedBook.set(UncraftEverythingConfig.outputEnchantedBook);
+        CONFIG.restrictAmbiguouslyCraftedItems.set(UncraftEverythingConfig.restrictAmbiguouslyCraftedItems);
         CONFIG_SPEC.save();
     }
 }