+++ changelog.md
@@ -1,1 +1,4 @@
-- Port to 26.2-snapshot-5
+# 26.2.0.0
+### Changes
+- Added new config field `Restrict Ambiguously Crafted Items` to common config
+- Recipe that has ItemTags ingredient can be disabled by setting `Restrict Ambiguously Crafted Items` to `false`